+++ fest/src/dataset_mul.c
@@ -109,7 +109,7 @@ int getDimensions(FILE* fp, int* examples, int* totalfeatures){
     return max+4; /* Just in case I was sloppy */
 }
 
-int readExamples(FILE* fp, int maxline, evpair_t* em, int* fm, int* trg){
+int readExamples(FILE* fp, int maxline, evpair_t* em, int* fm, short* trg){
     int i,target,example;
     float val;
     char* line;
@@ -153,7 +153,7 @@ int readExamples(FILE* fp, int maxline, evpair_t* em, int* fm, int* trg){
     return i;
 }
 
-int readExample(FILE* fp, int maxline, float* example, int nfeat, int* target){
+int readExample(FILE* fp, int maxline, float* example, int nfeat, short* target){
     int offset,feat,len;
     float val;
     char* line;
@@ -167,7 +167,7 @@ int readExample(FILE* fp, int maxline, float* example, int nfeat, int* target){
         comment=strchr(line,'#');
         if(comment!=NULL)
             *comment = '\0';
-        if(sscanf(line,"%d%n",target,&len)==EOF)
+        if(sscanf(line,"%hd%n",target,&len)==EOF)
             /* The line was a comment */
             continue;
         // *target = *target <=0 ? 0 : 1;  //  ********** binary classifictiation
@@ -222,8 +222,8 @@ void loadData(const char* name, dataset_t* d){
     }
     maxline=getDimensions(fp, &d->nex, &total);
 
-    d->target=malloc(d->nex*sizeof(int));
-    d->realtarget=malloc(d->nex*sizeof(int));
+    d->target=malloc(d->nex*sizeof(short));
+    d->realtarget=malloc(d->nex*sizeof(short));
     d->oobvotes=calloc(d->nex,sizeof(int));
     d->weight=malloc(d->nex*sizeof(float));
